Remove null callback on skipped info queries.
The use of null was conflicting with _queryInfo returning null if no
reddit info was found. Since no code needs to act on skipped requests,
omitting the callback makes sense. If we need to handle that case in the
future, we should probably use a special value instead of null.